Better Refactor. board vs board_algorithms
